+++ platformio/builder/tools/piomisc.py
@@ -288,10 +288,15 @@ def PioClean(env, clean_dir):
 def ProcessDebug(env):
     if not env.subst("$PIODEBUGFLAGS"):
         env.Replace(PIODEBUGFLAGS=["-Og", "-g3", "-ggdb3"])
-    env.Append(PIODEBUGFLAGS=["-D__PLATFORMIO_DEBUG__"])
     env.Append(
-        BUILD_FLAGS=env.get("PIODEBUGFLAGS", []),
-        BUILD_UNFLAGS=["-Os", "-O0", "-O1", "-O2", "-O3"])
+        PIODEBUGFLAGS=["-D__PLATFORMIO_DEBUG__"],
+        BUILD_FLAGS=env.get("PIODEBUGFLAGS", [])
+    )
+    unflags = ["-Os"]
+    for level in [0, 1, 2]:
+        for flag in ("O", "g", "ggdb"):
+            unflags.append("-%s%d" % (flag, level))
+    env.Append(BUILD_UNFLAGS=unflags)
 
 
 def ProcessTest(env):