+++ platformio/builder/tools/platformio.py
@@ -324,6 +324,7 @@ def BuildFrameworks(env, frameworks):
 
 
 def BuildLibrary(env, variant_dir, src_dir, src_filter=None):
+    env.ProcessUnFlags(env.get("BUILD_UNFLAGS"))
     return env.StaticLibrary(
         env.subst(variant_dir),
         env.CollectBuildFiles(variant_dir, src_dir, src_filter))