+++ HISTORY.rst
@@ -13,11 +13,12 @@ PlatformIO 3.0
 * Introduced `"Release" and "Debug" Build Configurations <http://docs.platformio.org/page/projectconf/build_configurations.html>`__
 * Build project in "Debug Mode" including debug information with a new ``debug`` target using `platformio run <https://docs.platformio.org/page/userguide/cmd_run.html>`__ command or `targets <http://docs.platformio.org/page/projectconf/section_env_general.html#targets>`__ option in ``platformio.ini``. The last option allows to avoid project rebuilding between "Run/Debug" modes.
   (`issue #1833 <https://github.com/platformio/platformio-core/issues/1833>`_)
-* Do not re-create ".gitignore" and ".travis.yml" files if they were removed from a project
+* Process ``build_unflags`` for cloned environment when building a static library
 * Report about outdated `99-platformio-udev.rules <http://docs.platformio.org/page/faq.html#platformio-udev-rules>`__
   (`issue #1823 <https://github.com/platformio/platformio-core/issues/1823>`_)
 * Show a valid error when Internet is off-line while initializing a new project
   (`issue #1784 <https://github.com/platformio/platformio-core/issues/1784>`_)
+* Do not re-create ".gitignore" and ".travis.yml" files if they were removed from a project
 * Fixed an issue when dynamic build flags were not handled correctly
   (`issue #1799 <https://github.com/platformio/platformio-core/issues/1799>`_)
 * Fixed an issue when ``pio run -t monitor`` always uses first ``monitor_port`` even with multiple environments